api/requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5  # PostgreSQL driver

# API Framework (assuming FastAPI based on project structure)
fastapi>=0.95.0
uvicorn>=0.22.0

# Settings and Environment
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Messaging and Caching
redis>=4.5.0
pika>=1.3.0  # For RabbitMQ

# Core dependencies
pandas>=1.3.0
numpy>=1.20.0
brotli>=1.0.9
torch>=1.10.0
uuid>=1.30.0

# LangChain dependencies
langchain>=0.0.267
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Embedding models
sentence-transformers>=2.2.2

# Document loaders
pypdf>=3.5.0

# Vector Database
weaviate-client>=3.15.0

# Utilities
requests>=2.28.0

# Añadir al final del archivo
python-multipart>=0.0.5

email-validator>=2.0.0